codegen: Add codegen for Views in Java
