+++ codegen/java-gen/src/test/scala/com/lightbend/akkasls/codegen/java/TestData.scala
@@ -1,0 +1,112 @@
+/*
+ * Copyright (c) Lightbend Inc. 2021
+ *
+ */
+
+package com.lightbend.akkasls.codegen
+package java
+
+object TestData {
+  def serviceProto(suffix: String = ""): PackageNaming =
+    PackageNaming(
+      s"MyService$suffix",
+      "com.example.service",
+      None,
+      None,
+      Some(s"ServiceOuterClass$suffix"),
+      javaMultipleFiles = false
+    )
+
+  def domainProto(suffix: String = ""): PackageNaming =
+    PackageNaming(
+      s"Domain$suffix",
+      "com.example.service.persistence",
+      None,
+      None,
+      Some(s"EntityOuterClass$suffix"),
+      javaMultipleFiles = false
+    )
+
+  val externalProto: PackageNaming =
+    PackageNaming(
+      "EXT",
+      "com.external",
+      None,
+      None,
+      None,
+      javaMultipleFiles = true
+    )
+
+  def simpleEntityService(
+      proto: PackageNaming = serviceProto(),
+      suffix: String = ""
+  ): ModelBuilder.EntityService =
+    ModelBuilder.EntityService(
+      FullyQualifiedName(s"MyService$suffix", proto),
+      List(
+        ModelBuilder.Command(
+          FullyQualifiedName("Set", proto),
+          FullyQualifiedName("SetValue", proto),
+          FullyQualifiedName("Empty", externalProto)
+        ),
+        ModelBuilder.Command(
+          FullyQualifiedName("Get", proto),
+          FullyQualifiedName("GetValue", proto),
+          FullyQualifiedName("MyState", proto)
+        )
+      ),
+      s"com.example.Entity$suffix"
+    )
+
+  def simpleViewService(
+      proto: PackageNaming = serviceProto(),
+      suffix: String = ""
+  ): ModelBuilder.ViewService =
+    ModelBuilder.ViewService(
+      FullyQualifiedName(s"MyService$suffix", proto),
+      List(
+        ModelBuilder.Command(
+          FullyQualifiedName("Created", proto),
+          FullyQualifiedName("EntityCreated", proto),
+          FullyQualifiedName("ViewState", proto)
+        ),
+        ModelBuilder.Command(
+          FullyQualifiedName("Updated", proto),
+          FullyQualifiedName("EntityUpdated", proto),
+          FullyQualifiedName("ViewState", proto)
+        )
+      ),
+      List(
+        ModelBuilder.Command(
+          FullyQualifiedName("Created", proto),
+          FullyQualifiedName("EntityCreated", proto),
+          FullyQualifiedName("ViewState", proto)
+        ),
+        ModelBuilder.Command(
+          FullyQualifiedName("Updated", proto),
+          FullyQualifiedName("EntityUpdated", proto),
+          FullyQualifiedName("ViewState", proto)
+        )
+      ),
+      List.empty
+    )
+
+  def eventSourcedEntity(
+      suffix: String = ""
+  ): ModelBuilder.EventSourcedEntity =
+    ModelBuilder.EventSourcedEntity(
+      FullyQualifiedName(s"MyEntity$suffix", domainProto(suffix)),
+      s"MyEntity$suffix",
+      Some(ModelBuilder.State(FullyQualifiedName("MyState", domainProto(suffix)))),
+      List(
+        ModelBuilder.Event(FullyQualifiedName("SetEvent", domainProto(suffix)))
+      )
+    )
+
+  def valueEntity(suffix: String = ""): ModelBuilder.ValueEntity =
+    ModelBuilder.ValueEntity(
+      FullyQualifiedName(s"MyValueEntity$suffix", domainProto(suffix)),
+      s"MyValueEntity$suffix",
+      ModelBuilder.State(FullyQualifiedName("MyState", domainProto(suffix)))
+    )
+}
